proxy mode for controller
proxy mode allows a user connected to the controller via USB to open a direct proxy channel to the remote bus: the controller reads incoming data (from USB) and forwards it to the busses. IOIB data is ignored and no refreshing is performed: the controller becomes a dumb proxy until it is reset.